inisial commit

+++ backend/controllers/goalControllers.js
@@ -1,10 +1,11 @@
 const asyncHandler= require('express-async-handler')
 const Goal   = require('../moduls/goalModul')
+const User   = require('../moduls/useModul')
 
 
 
 const getGoals = asyncHandler(async(req, res ) =>{
-    const goals = await Goal.find ()
+    const goals = await Goal.find ({user : req.user.id})
     res.json(goals)
 
 })
@@ -14,7 +15,8 @@ const setGoals = asyncHandler(async(req, res ) =>{
     throw new Error("please add text fiald")
    } 
    const goal = await Goal.create({
-    text : req.body.text
+    text : req.body.text,
+    user:req.user.id,
    })
     res.json(goal)
 
@@ -26,6 +28,19 @@ const UpdateGoals = asyncHandler(async(req,res)=>{
         res.status(400)
         throw new Error('Goal not found ')
     }
+    const user = await User.findById(req.user.id)
+    //check for users
+    if (!user ){
+        res.status(401)
+        throw new Error('not found')
+    }
+    //make sure the logged in user matches the goal user
+    if (goal.user.toString()!= user.id){
+        res.status(401)
+        throw new Error('user no authorized')
+    }
+
+
     const updatedGoal= await Goal.findByIdAndUpdate(req.params.id ,req.body ,{new : true})
     res.json(updatedGoal)
 })
@@ -35,6 +50,18 @@ const deleteGoals = asyncHandler(async(req,res)=>{
     if(!goal){
         res.status(400)
         throw new Error('Goal not found ')
+    }
+
+    const user = await User.findById(req.user.id)
+    //check for users
+    if (!user ){
+        res.status(401)
+        throw new Error('not found')
+    }
+    //make sure the logged in user matches the goal user
+    if (goal.user.toString()!= user.id){
+        res.status(401)
+        throw new Error('user no authorized')
     }
    await goal.deleteOne();
     res.json({id : req.params.id})
